Intent: Preserve gamma coverage during suite consolidation Operators involved: [Coherence] Affected invariants: [#1] Key changes: - Port the gamma none evaluation check into tests/test_gamma.py - Remove duplicated gamma variant cases and drop their module Expected risks/dissonances: - None; tests ensure the consolidated suite behaves as before Metrics: - N/A Equivalence map: - tests/test_gamma_variants.py coverage is subsumed by tests/test_gamma.py
